Move PacmanMenu to archinstall.lib.pacman package
